nasa_images_fetcher/requirements.txt
earthengine-api==0.1.357
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0